Quiz_week10/requirements.txt
Flask
Flask-Migrate
flask-marshmallow
Flask-SQLAlchemy
PyMySQL
requests
beautifulsoup4
lxml